feat/3-Axios-Debt-Types: * added generic function to map arrays * refactored implementations that could the array mapper * created query key constant to hold all the query keys used my tanstack, refactored previous implementations to use this constant
