+++ src/layouts/debt/components/AddDebtForm/index.js
@@ -1,17 +1,24 @@
 import MDBox from "../../../../components/MDBox";
 import MDInput from "../../../../components/MDInput";
 import MDButton from "../../../../components/MDButton";
-import { addDebt } from "../../../../api/user";
+import { addDebt, getDebtTypes } from "../../../../api/user";
 import MDTypography from "../../../../components/MDTypography";
-import { useMutation, useQueryClient } from "@tanstack/react-query";
+import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
+import { QUERY_KEY } from "../../../../utils/consts/queryKeys";
+import arrayMapper from "../../../../utils/arrays/arrayMapper";
 
 export function AddDebtForm() {
   const queryClient = useQueryClient();
 
+  const { data: debtTypes } = useQuery({
+    queryKey: QUERY_KEY.GET_DEBT_TYPE,
+    queryFn: async () => arrayMapper(await getDebtTypes(), "type"),
+  });
+  console.log(`Debt Types: ${debtTypes || []}`);
+
   const mutation = useMutation({
     mutationFn: (e) => addDebt(e),
     onSuccess: () => {
-      debugger;
       queryClient.invalidateQueries();
     },
   });
