+++ src/utils/arrays/arrayMapper.js
@@ -1,0 +1,3 @@
+export default function arrayMapper(objectArray, mapping) {
+  return objectArray.map((item) => item[mapping]);
+}
