+++ src/utils/charts/formattedLineChartData.js
@@ -1,7 +1,9 @@
+import arrayMapper from "../arrays/arrayMapper";
+
 export default function formattedLineChartData(data) {
   debugger;
   return {
-    labels: data?.map((item) => item.paymentDate) || [],
-    datasets: { label: "Payment", data: data?.map((item) => item.payment) || [] },
+    labels: arrayMapper(data, "paymentDate") || [],
+    datasets: { label: "Payment", data: arrayMapper(data, "payment") || [] },
   };
 }
